requirements.txt
dvc==3.55.2
dvc-data==3.16.5
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.4.0
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
grpcio==1.66.1
gto==1.7.1
keras==2.10.0
Keras-Preprocessing==1.1.2
mysql-connector-python==9.0.0
numpy==1.21.6
pillow==10.4.0
PyYAML==6.0.2
protobuf>=3.9.2,<3.20
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.31.0
ipython==8.27.0
